api/requirements.txt
fastapi
uvicorn
python-multipart
realitydefender
numpy
scikit-learn
pillow
opencv-python-headless
scipy
python-dotenv
